README.md 파일 및 사칙연산 문제 풀이

+++ boj25q/src/Arithmetic/boj1008.java
@@ -10,7 +10,7 @@ public class boj1008 {
         int A = sc.nextInt();
         int B = sc.nextInt();
 
-        System.out.println(A / B);
+        System.out.println((double)A / B);
 
         sc.close();
     }
